Documentation: ditaa: pick command invocation from env vars
Currently the ditaa command is invoked directly and assumes that
binfmt_misc support is enabled on the host to execute ditaa via jexec
or jarwrapper. However that is not true in all cases (e.g. docker
guests does not yet properly support binfmt_misc).

This commit adds support for command invocation customization with two
environment variables: SPHINX_DITAA_CMD to specify the executable and
SPHINX_DITAA_ARG to specify a list of parameters.

Signed-off-by: Octavian Purdila <tavi@cs.pub.ro>