templates: 6-list-sync: Fix potential dangling pointer
As soon as the lock is unlocked in task_info_find_pid,
the caller cannot be sure if the pointer that is returned is valid.
As long as the returned pointer is used, the lock must be held.

In the case of task_info_add_to_list, the lock that must be held is a writer lock,
in case of a reader lock other readers could read an inconsistent state of the struct task_info.