[5-PITIX] Add score information in checker
For 5-pitix assignment there is no info about score or the number of test passed (at least a counter). Simple searching for "failed" in output is error prone (and painfully hard for the average Joe).

A simple solution is to add a test_ok var to count the number of passed test. Better solutions exist, but keep it simple stupid as fellow student only care about the number of tests passed to know the homework is okay.