Documentation/filesystems_part1/minfs: Fix wrong target block device
The document instructs to run 'mkfs.minfs' for '/dev/vdb', but because
we have 2 additional disks already and we added 'mydisk.img' in the
beginning of the exercise, '/dev/vdd' must be the one we want to use.
Also, the mount instruction points the device with '/dev/vdX'.

This commit updates the doc for this.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>