tools: labs: Bring back CONFIG_64BIT is not set
Using ARCH=x86 unifies i386 and x86_64 architecture
and lets the build system to decide between 32/64 bits
based on .config file.

Commit 86b869d0f1d351 ("tools: labs: Enable VIRTO_PCI")
removed this line:

   -# CONFIG_64BIT is not set

which apparently it is very important. It decides 32 vs 64 bit
compilation.

Bring it back so that with our default kernel_config.x86 config
we will always compile for 32 bits.

Signed-off-by: Daniel Baluta <daniel.baluta@nxp.com>