sphinx: asciicast: install .cast files to _images
Arguably _images is a better place to store the .cast files. It also
fixes the following build issues:

Exception occurred:
  File "/usr/lib/python3.8/shutil.py", line 261, in copyfile
    with open(src, 'rb') as fsrc, open(dst, 'wb') as fdst:
FileNotFoundError: [Errno 2] No such file or directory: Documentation/output/teaching/lectures/file.cast

The issue here is that asciicast "rendering" happens before html
rendering and because of that the output lectures directory is not yet
available.

Signed-off-by: Octavian Purdila <tavi@cs.pub.ro>