labs: interrupts: Fix release of DATA_REG
In the touched code we are trying to request two regions. But in the
case when the first request would succeed and the second one fail, we
would not correctly free the first region on the error path.

This commit fixes that by introducing an additional label and changing
a goto destination.