Solve `make console` bugs for docker
In this commit, I added a rule to install samba in the
docker because it is required by the new setup. And I
removed the qemu_kvm variable from the qemu/run-qemu.sh
script when kvm is not available (It is easier this way
than setting the QEMU_KVM environment variable).

Signed-off-by: Dragos Iulian Argint <dragosargint21@gmail.com>