Add support for taps and disks to run-qemu.sh
The New version of running the virtual machine
(the one that uses `make console`) lacks the network
interfaces and disks that will be used for homework
and labs. This commit adds support for the 2 network
interferences, plus 1 more that is used by the samba
server, and the 2 virtual disks. All of them can be
found in the old setup (the one that uses `make boot`).

Signed-off-by: Dragos Iulian Argint <dragosargint21@gmail.com>